Re-structured the project layout, minimap separated
