+++ includes/checkers.h
@@ -1,0 +1,8 @@
+#ifndef CHECKERS_H
+# define CHECKERS_H
+
+#include "types.h"
+
+bool	is_prog_valid(t_prog *prog);
+
+#endif
