+++ includes/maths.h
@@ -1,0 +1,7 @@
+#ifndef MATHS_H
+# define MATHS_H
+
+
+int calculate_distance(int px, int py, int dx, int dy);
+
+#endif
