+++ includes/minimap_utils.h
@@ -1,0 +1,7 @@
+#ifndef MINIMAP_UTILS_H
+# define MINIMAP_UTILS_H
+
+#include "types.h"
+int	minimap(t_prog *prog);
+
+#endif
