+++ includes/mlx_draw.h
@@ -6,5 +6,6 @@
 int mlx_draw_minimap(t_mlx *mlx, t_info *info);
 bool	mlx_draw_field(t_mlx *mlx, t_info *info);
 bool	mlx_draw_square(t_mlx *mlx, int x, int y, int color, int size);
-
+int	draw_line(t_mlx *mlx, t_info *info, int x, int y, double dx, double dy, int color);
+void	put_pixel(t_mlx *mlx, int x, int y, int color);
 #endif
