+++ includes/mlx_game_simulation.h
@@ -2,4 +2,5 @@
 # define MLX_GAME_SIMULATION_H
 #include "types.h"
 bool	mlx_simulate_game(t_prog *prog);
+int	game_layout(t_prog *prog);
 #endif
